feat(write_draft, write_introduction): use a single folder and only call in certain files
automate the creation of some files, the code is not generalized,
just for lingcod right now.

committing initial model files for preamble stuff
deleted doc\lingcod_21f_rec_00.R b/c using new structure

+++ R/write_introduction.R
@@ -18,7 +18,10 @@
 #' @export
 #'
 write_introduction <- function(dir, dirmaster = "..") {
-  unlink(file.path(dir, "11introduction.Rmd"))
+  if (file.exists(file.path(dir, "11introduction.Rmd"))) {
+    unlink(file.path(dir, "11introduction.Rmd"))
+  }
+
   writeLines(
     con = file.path(dir, "11introduction.Rmd"),
     c(
@@ -31,7 +34,7 @@ write_introduction <- function(dir, dirmaster = "..") {
         dirmaster,
         "', pattern = '^1[2-9][a-z]*\\\\.Rmd', full.names = TRUE)"
       ),
-      "ignore <- mapply(read_child, files_intro)",
+      "ignore <- mapply(sa4ss::read_child, files_intro)",
       "```\n"
     )
   )
